Fix build on g++-4.x.

+++ src/contrib/flow-monitor/histogram.cc
@@ -78,7 +78,7 @@ Histogram::GetBinCount (uint32_t index)
 void 
 Histogram::AddValue (double value)
 {
-  uint32_t index = floor (value/m_binWidth);
+  uint32_t index = (uint32_t)floor (value/m_binWidth);
   
   //check if we need to resize the vector
   NS_LOG_DEBUG ("AddValue: index=" << index << ", m_histogram.size()=" << m_histogram.size ());
